requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

chainlit==0.7.700
langchain==0.0.322
python-dotenv==1.0.0
autogen==1.0.16
pyautogen==0.2.0b6
